--- core/src/main/java/nl/nn/adapterframework/jdbc/dbms/OracleKeyWords.java
@@ -1,39 +1,0 @@
-/*
-   Copyright 2020 WeAreFrank!
-
-   Licensed under the Apache License, Version 2.0 (the "License");
-   you may not use this file except in compliance with the License.
-   You may obtain a copy of the License at
-
-       http://www.apache.org/licenses/LICENSE-2.0
-
-   Unless required by applicable law or agreed to in writing, software
-   distributed under the License is distributed on an "AS IS" BASIS,
-   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-   See the License for the specific language governing permissions and
-   limitations under the License.
-*/
-package nl.nn.adapterframework.jdbc.dbms;
-
-public enum OracleKeyWords {
-
-	NEXTVAL(".NEXTVAL"),
-	CURRVAL(".CURRVAL"),
-	FOR_UPDATE("FOR UPDATE"),
-	EMPTY_CLOB("EMPTY_CLOB"),
-	SYSDATE("SYSDATE"),
-	SYSTIMESTAMP("SYSTIMESTAMP"),
-	EMPTY_BLOB("EMPTY_BLOB"),
-	DUAL("DUAL");
-	
-	private final String key;
-	
-	OracleKeyWords(String key) {
-		this.key = key;
-	}
-	
-	public String key() {
-		return this.key;
-	}
-	
-}
