+++ docker/pom.xml
@@ -17,7 +17,6 @@
 	<properties>
 		<iaf.version>7.7-SNAPSHOT</iaf.version>
 		<revision>${iaf.version}</revision>
-		<h2.version>2.0.206</h2.version>
 		<oracle.driver.groupid>com.oracle.database.jdbc</oracle.driver.groupid>
 		<oracle.driver.version>18.3.0.0</oracle.driver.version>
 		<oracle.driver.jdkversion>8</oracle.driver.jdkversion>
