+++ example/pom.xml
@@ -29,7 +29,7 @@
 		<dependency>
 			<groupId>com.h2database</groupId>
 			<artifactId>h2</artifactId>
-			<version>2.0.206</version>
+			<version>${h2.version}</version>
 		</dependency>
 		<!-- Includes missing javax.jms.JMSExceptions class -->
 		<dependency>
